UV Basics/UV_Litellm/pyproject.toml
[project]
name = "uv-litellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "legacy-cgi>=2.6.2",
    "litellm>=1.59.10",
    "python-dotenv>=1.0.1",
]
